fix: remove hydration
