+++ src/app/developers/page.tsx
@@ -7,44 +7,46 @@ const developers = [
     id: 1,
     name: "山田 太郎",
     role: "リードエンジニア",
-    description: "フロントエンド開発のリーダー。ReactとTypeScriptのスペシャリスト。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=taro"
+    description:
+      "フロントエンド開発のリーダー。ReactとTypeScriptのスペシャリスト。",
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=taro",
   },
   {
     id: 2,
     name: "佐藤 花子",
     role: "UIデザイナー",
-    description: "ユーザーインターフェースとユーザーエクスペリエンスのデザインを担当。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=hanako"
+    description:
+      "ユーザーインターフェースとユーザーエクスペリエンスのデザインを担当。",
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=hanako",
   },
   {
     id: 3,
     name: "鈴木 一郎",
     role: "バックエンドエンジニア",
     description: "サーバーサイドの開発とデータベース設計を担当。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=ichiro"
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=ichiro",
   },
   {
     id: 4,
     name: "高橋 雄太",
     role: "プロジェクトマネージャー",
     description: "プロジェクト全体の進行管理と調整を担当。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=yuta"
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=yuta",
   },
   {
     id: 5,
     name: "伊藤 美咲",
     role: "QAエンジニア",
     description: "品質保証とテスト自動化を担当。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=misaki"
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=misaki",
   },
   {
     id: 6,
     name: "渡辺 健太",
     role: "DevOpsエンジニア",
     description: "CI/CDパイプラインとインフラストラクチャの管理を担当。",
-    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=kenta"
-  }
+    avatar: "https://api.dicebear.com/7.x/avataaars/svg?seed=kenta",
+  },
 ];
 
 interface Developer {
@@ -70,18 +72,14 @@ const DeveloperCard = ({ developer }: { developer: Developer }) => {
       }}
     >
       <Flex direction="column" align="center">
-        <Avatar
-          src={developer.avatar}
-          size="xl"
-          mb={4}
-        />
+        <Avatar src={developer.avatar} size="xl" mb={4} />
         <Heading as="h3" size="md" mb={2} textAlign="center">
           {developer.name}
         </Heading>
-        <Text 
-          fontSize="sm" 
-          fontWeight="bold" 
-          color="gray.500" 
+        <Text
+          fontSize="sm"
+          fontWeight="bold"
+          color="gray.500"
           _dark={{ color: "gray.400" }}
           mb={3}
         >
@@ -108,11 +106,11 @@ export default function Developers() {
         templateColumns={{
           base: "repeat(1, 1fr)",
           md: "repeat(2, 1fr)",
-          lg: "repeat(3, 1fr)"
+          lg: "repeat(3, 1fr)",
         }}
         gap={6}
       >
-        {developers.map(developer => (
+        {developers.map((developer) => (
           <DeveloperCard key={developer.id} developer={developer} />
         ))}
       </Grid>
