+++ src/components/mode-change/index.tsx
@@ -1,16 +1,10 @@
 "use client";
+
 import { useDarkLight } from "@/libs/dark-light";
 import { SunIcon } from "@yamada-ui/lucide";
-import { useEffect, useState } from "react";
 
 export function ModeChange() {
   const { state, changeMode } = useDarkLight();
-  const [mounted, setMounted] = useState(false);
-
-  // Fix hydration mismatch by only applying styles after client-side hydration
-  useEffect(() => {
-    setMounted(true);
-  }, []);
 
   const handleClick = () => {
     if (state.type === "dark") {
@@ -19,11 +13,6 @@ export function ModeChange() {
       changeMode("dark");
     }
   };
-
-  // Don't render the button until after client-side hydration
-  if (!mounted) {
-    return null;
-  }
 
   return (
     <SunIcon
