+++ src/libs/dark-light/store/dark-light.type.ts
@@ -1,13 +1,15 @@
+import { DARK_BG, DARK_COLOR, LIGHT_BG, LIGHT_COLOR } from "@/theme";
+
 interface Dark {
   type: "dark";
-  bg: "#32373c"; 
-  color: "#E8E6E3"; 
+  bg: typeof DARK_BG;
+  color: typeof DARK_COLOR;
 }
 
 interface Light {
   type: "light";
-  bg: "#F2F1EE"; 
-  color: "#32373c"; 
+  bg: typeof LIGHT_BG;
+  color: typeof LIGHT_COLOR;
 }
 
 export type ModeState = Dark | Light;
