+++ src/theme/index.ts
@@ -1,7 +1,6 @@
-"use client";
 import { extendTheme } from "@yamada-ui/react";
 
-export const theme = extendTheme({
+export const yamadaTheme = extendTheme({
   styles: {
     global: {
       body: {
@@ -65,3 +64,8 @@ export const theme = extendTheme({
     },
   },
 });
+
+export const DARK_BG = "#32373c" as const;
+export const LIGHT_BG = "#F2F1EE" as const;
+export const DARK_COLOR = "#E8E6E3" as const;
+export const LIGHT_COLOR = "#32373c" as const;
